Block Segment copy constructor and assignment operator, use explicit take/copy
Having a copy constructor (which shallow copied) and an assignment
operator (which deep copied) was non-obvious and dangerous, because
it was too easy to generate implicit shallow or deep copies using
transient Segments.  The first is dangerous because of aliasing,
the second inefficient.

Now both are blocked by being made private and there is are copy()
(deep) and take() (shallow, clearing the source) operations.  The
rest of the library has been changed to use them.