Provide ConstXPathProcessor using Deep Magic template trickery
BaseXPathProcessor is a template that can be specialised with either
Element or const Element.  By using explicit specialisation we can
keep the implementation of this in xpath.cc.

ConstXPathProcessor is just the specialisation of this to const

XPathProcessor specialises to non-const and adds the set methods,
providing the identical interface to before.