Pass and store optional subscriber IDs with messages
Publisher now stores and reflects a subscriber ID back with a subscribed
message, and HTTP server uses this to split the concept of the specific
connection path (used for request responses) and subscriber_id (used for
polling).

This means that subscription responses are returned properly to their
connections, but pollers can still receive subscribed messages on a separate
connection.

bug:fixed:808