Change Time::Stamp internal format to solve post-2036
NTP timestamp rolls over in 2036, and hence we couldn't previously
represent anything beyond that.  Changing the internal fixed point to 20
instead of 32 bits still allows us microsecond resolution, but gains us
half a million years.

If someone is reading this in half a million years time on a 64-bit
computer, sorry.