Revert to temporary certs
OpenSSL is inconsistent - SSL_CTX_use_certificate seems to either
copy the cert or up its reference count, so it's safe to pass in
a temporary one which we will free again.

SSL_CTX_add_extra_chain_cert will free it itself, so we need to
up the refcount before passing it to them.

Bletch (and thanks for a wasted day!)